requirements.txt
async-timeout==4.0.2
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
Flask==2.2.2
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
MouseInfo==0.1.3
packaging==21.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dotenv==0.21.0
pytweening==1.0.4
redis==4.3.4
requests==2.28.1
six==1.16.0
spotipy==2.21.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
